debt: Remove HelloWorks/Dropbox Forms
