+++ docs/environment_variables.md
@@ -41,11 +41,8 @@
 | FOREST_AUTH_SECRET                          |                                                    | forest auth secret                                                                   |
 | FOREST_ENV_SECRET                           |                                                    | forest environment secret                                                            |
 |                                             | .stripe.redirectUri                                |                                                                                      |
-| HELLO_WORKS_KEY                             | .helloworks.key                                    | HelloWorks key                                                                       |
-| HELLO_WORKS_SECRET                          | .helloworks.secret                                 | HelloWorks secret                                                                    |
 | TAX_FORMS_ENCRYPTION_KEY                    | .taxForms.encryptionKey                            | base64 encoded secret key for encrypting document before storage.                    |
 | TAX_FORMS_AWS_S3_BUCKET                     | .taxforms.aws.s3.bucket                            | the bucket where tax forms will be uploaded                                          |
-| TAX_FORMS_USE_INTERNAL                      | .taxForms.useInternal                              | whether to use the internal tax forms system                                         |
 | GITHUB_FLOW_MIN_NB_STARS                    | .githubFlow.minNbStars                             | Minimum number of Github stars required to apply to the open source collective       |
 | TRANSFERWISE_API_URL                        | .transferwise.apiUrl                               |                                                                                      |
 | TRANSFERWISE_OAUTH_URL                      | .transferwise.oauthUrl                             |                                                                                      |
