+++ package.json
@@ -89,7 +89,6 @@
     "handlebars": "4.7.8",
     "hashids": "2.3.0",
     "hcaptcha": "0.1.1",
-    "helloworks-sdk": "1.3.0",
     "helmet": "7.1.0",
     "html-pdf": "3.0.1",
     "html-to-text": "9.0.5",
@@ -109,7 +108,6 @@
     "p-filter": "2.1.0",
     "p-map": "4.0.0",
     "p-queue": "6.6.2",
-    "p-throttle": "4.1.1",
     "passport": "0.7.0",
     "passport-github": "1.1.0",
     "paypal-adaptive": "0.4.2",
