+++ server/lib/notifications/email.ts
@@ -11,7 +11,6 @@ import { TransactionKind } from '../../constants/transaction-kind';
 import { TransactionTypes } from '../../constants/transactions';
 import models, { Activity, Collective } from '../../models';
 import { CommentType } from '../../models/Comment';
-import { LEGAL_DOCUMENT_SERVICE } from '../../models/LegalDocument';
 import { UpdateChannel } from '../../models/Update';
 import User from '../../models/User';
 import emailLib from '../email';
@@ -208,11 +207,6 @@ export const notifyByEmail = async (activity: Activity) => {
     case ActivityTypes.ORGANIZATION_COLLECTIVE_CREATED:
     case ActivityTypes.USER_CARD_CLAIMED:
       await notify.user(activity);
-      break;
-
-    // Send tax form requests emails only for Dropbox form, others are sent as reminders in cron/hourly/40-send-tax-form-requests.js
-    case ActivityTypes.TAXFORM_REQUEST:
-      activity.data?.service === LEGAL_DOCUMENT_SERVICE.DROPBOX_FORMS && (await notify.user(activity));
       break;
 
     case ActivityTypes.TAXFORM_INVALIDATED:
